refactor(angular-base): refactored and seperated prop and method statements
