+++ packages/teleport-plugin-angular-base-component/src/utils.ts
@@ -2,7 +2,7 @@ import * as types from '@babel/types'
 import {
   convertValueToLiteral,
   getTSAnnotationForType,
-  createMethodsObject,
+  createStateChangeStatement,
 } from '@teleporthq/teleport-shared/dist/cjs/utils/ast-js-utils'
 import {
   UIDLPropDefinition,
@@ -20,15 +20,12 @@ export const generateExportAST = (
 ) => {
   let angularMethodsAST = []
   if (Object.keys(methodsObject).length > 0) {
-    angularMethodsAST = createMethodsObject(
-      methodsObject,
-      propDefinitions,
-      'angular'
-    ) as types.ClassMethod[]
+    angularMethodsAST = createMethodsObject(methodsObject, propDefinitions)
   }
 
   const propDeclaration = Object.keys(propDefinitions).map((propKey) => {
     const definition = propDefinitions[propKey]
+    // By default any prop with type function is used to emitting events to the callee
     if (definition.type === 'func') {
       return t.classProperty(
         t.identifier(propKey),
@@ -85,4 +82,55 @@ export const generateExportAST = (
 
 const constructorAST = (t = types) => {
   return t.classMethod('constructor', t.identifier('constructor'), [], t.blockStatement([]))
+}
+
+const createMethodsObject = (
+  methods: Record<string, UIDLEventHandlerStatement[]>,
+  propDefinitions: Record<string, UIDLPropDefinition>,
+  t = types
+) => {
+  return Object.keys(methods).map((eventKey) => {
+    const astStatements = []
+    methods[eventKey].map((statement) => {
+      const astStatement =
+        statement.type === 'propCall'
+          ? createPropCallStatement(statement, propDefinitions)
+          : createStateChangeStatement(statement)
+
+      if (astStatement) {
+        astStatements.push(astStatement)
+      }
+    })
+    return t.classMethod('method', t.identifier(eventKey), [], t.blockStatement(astStatements))
+  })
+}
+
+export const createPropCallStatement = (
+  eventHandlerStatement: UIDLEventHandlerStatement,
+  propDefinitions: Record<string, UIDLPropDefinition>,
+  t = types
+) => {
+  const { calls: propFunctionKey } = eventHandlerStatement
+
+  if (!propFunctionKey) {
+    console.warn(`No prop definition referenced under the "calls" field`)
+    return null
+  }
+
+  const propDefinition = propDefinitions[propFunctionKey]
+
+  if (!propDefinition) {
+    console.warn(`No prop definition was found for function "${propFunctionKey}"`)
+    return null
+  }
+
+  return t.expressionStatement(
+    t.callExpression(
+      t.memberExpression(
+        t.memberExpression(t.thisExpression(), t.identifier(propFunctionKey)),
+        t.identifier('emit')
+      ),
+      []
+    )
+  )
 }
