+++ packages/teleport-shared/src/node-handlers/node-to-html/constants.ts
@@ -6,7 +6,7 @@ export const DEFAULT_TEMPLATE_SYNTAX: HTMLTemplateSyntax = {
   eventBinding: (value) => `@${value}`,
   valueBinding: (value) => `:${value}`,
   eventEmmitter: (value) => `this.$emit('${value}')`,
-  eventNameBinding: (value) => `${value}`,
+  eventHandlersBindingMode: (value) => value,
   conditionalAttr: 'v-if',
   repeatAttr: 'v-for',
   repeatIterator: (iteratorName, iteratedCollection, useIndex) => {
